requirements.txt
MetaTrader5>=5.0.33
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
xgboost>=1.7.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
